requirements.txt
Flask==3.1.2
openai==0.28.1
python-dotenv==1.1.1
